// client/src/app/shared/models/plant-details.ts
export type CareLevel = 'Hight' | 'Moderate' | 'Medium';
export type Watering = 'Low' | 'Medium' | 'High';
export type Sunlight =
  | 'Full Sun'
  | 'Partial Sun'
  | 'Indirect Light'
  | 'Low Light';

// Standardized Watering Options:

// "Low": Water every 2-3 weeks
// "Medium": Water once a week
// "High": Water multiple times a week
// Standardized Sunlight Options:

// "Full Sun"
// "Partial Sun"
// "Indirect Light"
// "Low Light"